изменено:      Android/BluZ_LPM/.idea/deploymentTargetSelector.xml 	удалено:       Android/BluZ_LPM/app/build/outputs/apk/release/app-release.apk 	изменено:      Android/BluZ_LPM/app/src/main/AndroidManifest.xml 	изменено:      Android/BluZ_LPM/app/src/main/java/ru/starline/bluz/BluetoothInterface.kt 	изменено:      Android/BluZ_LPM/app/src/main/java/ru/starline/bluz/MainActivity.kt 	изменено:      Android/BluZ_LPM/app/src/main/java/ru/starline/bluz/SaveBqMon.kt 	изменено:      Android/BluZ_LPM/app/src/main/java/ru/starline/bluz/globalObj.kt 	изменено:      Android/BluZ_LPM/app/src/main/java/ru/starline/bluz/intervalTimer.kt 	изменено:      Android/BluZ_LPM/app/src/main/res/values/strings.xml

+++ Android/BluZ_LPM/app/src/main/java/ru/starline/bluz/intervalTimer.kt
@@ -27,14 +27,18 @@ class intervalTimer {
 
 internal class MyTimerTask : TimerTask() {
     override fun run() {
-        if (GO.needTerminate) {     // Завершение приложения
-            System.exit(0)
-        }
-        if (GO.BTT.connected) {
-            // BT connected
+        if (GO.allPermissionAccept) {
+            if (GO.needTerminate) {     // Завершение приложения
+                System.exit(0)
+            }
+            if (GO.BTT.connected) {
+                // BT connected
+            } else {
+                GO.BTT.destroyDevice()
+                GO.BTT.initLeDevice()
+            }
         } else {
-            GO.BTT.destroyDevice()
-            GO.BTT.initLeDevice()
+            GO.BTT.connected = false
         }
     }
 }
